separate sam_worker
