+++ evaluation-service/backend/server.py
@@ -10,176 +10,82 @@ from loguru import logger
 from werkzeug.utils import secure_filename
 
 sys.path.append("..")
+sys.path.append(".")
+
 import datetime
 import glob
 import json
 import os
 
-import cv2
 import numpy as np
-import torch
-from segment_anything import SamPredictor, sam_model_registry
+from sam_worker import sam_worker
 
 app = Flask(__name__)
 CORS(app)  #
 
 from multiprocessing import Process, Queue
 
-
-def euler_to_mat(roll, pitch, yaw):
-    # Create rotation matrix from Euler angles
-    R_x = np.array([[1, 0, 0], [0, np.cos(roll), -np.sin(roll)], [0, np.sin(roll), np.cos(roll)]])
-
-    R_y = np.array([[np.cos(pitch), 0, np.sin(pitch)], [0, 1, 0], [-np.sin(pitch), 0, np.cos(pitch)]])
-
-    R_z = np.array([[np.cos(yaw), -np.sin(yaw), 0], [np.sin(yaw), np.cos(yaw), 0], [0, 0, 1]])
-
-    R = np.dot(R_z, np.dot(R_y, R_x))
-
-    return R
+from airo_camera_toolkit.pinhole_operations.projection import project_points_to_image_plane
+from airo_dataset_tools.data_parsers.camera_intrinsics import CameraIntrinsics
+from airo_spatial_algebra import transform_points
 
 
-def sam_worker(q, ack, args):
+def get_heuristic_cloth_bounding_box(sample_dir: str) -> tuple[float, float, float, float]:
+    """Calculates an approximate 2D bounding box for the cloth region held by the robot arms.
 
-    sam_checkpoint = "../weights/sam_vit_b_01ec64.pth"
-    model_type = "vit_b"
-    device = "cpu"
+    This function assume the case where the cloth is held both robots arms and stretched in front of the camera.
 
-    cuda = args.cuda
-    model_size = args.model
-    scenes_directory = args.scenes_directory
+    Args:
+        sample_dir: The path to the sample directory containing the necessary data files.
 
-    if model_size == "h":
-        sam_checkpoint = "../weights/sam_vit_h_4b8939.pth"
-        model_type = "vit_h"
-    elif model_size == "l":
-        sam_checkpoint = "../weights/sam_vit_l_0b3195.pth"
-        model_type = "vit_l"
+    Returns:
+        A tuple of (u_min, v_min, u_max, v_max) representing the coordinates of the
+        estimated bounding box within the image.
+    """
+    # Load data
+    tcp_left = json.load(open(f"{sample_dir}/observation_result/arm_left_tcp_pose_in_world.json"))
+    tcp_right = json.load(open(f"{sample_dir}/observation_result/arm_right_tcp_pose_in_world.json"))
 
-    if cuda and torch.cuda.is_available():
-        logger.success("Segment-Anything worker running on GPU")
-        device = "cuda"
-    else:
-        logger.warning("Segment-Anything worker running on CPU")
+    intrinsics_path = f"{sample_dir}/observation_result/camera_intrinsics.json"
+    with open(intrinsics_path, "r") as f:
+        intrinsics_model = CameraIntrinsics.model_validate_json(f.read())
+        intrinsics = intrinsics_model.as_matrix()
 
-    sam = sam_model_registry[model_type](checkpoint=sam_checkpoint)
-    sam.to(device=device)
+    extrinsics_path = f"{sample_dir}/observation_result/camera_pose_in_world.json"
+    with open(extrinsics_path) as f:
+        extrinsics = Pose.model_validate_json(f.read()).as_homogeneous_matrix()
 
-    predictor = SamPredictor(sam)
+    tcp_left_position = tcp_left["position_in_meters"]
+    tcp_right_position = tcp_right["position_in_meters"]
 
-    while True:
-        print("Waiting for data")
-        data = q.get()
-        print("Data received", data)
+    x_left = tcp_left_position["x"]
+    y_left = tcp_left_position["y"]
+    z_left = tcp_left_position["z"]
 
-        manual = data["manual"]
+    x_right = tcp_right_position["x"]
+    y_right = tcp_right_position["y"]
+    z_right = tcp_right_position["z"]
 
-        bbox = data["bbox"]
-        positive_keypoints = data["positiveKeypoints"]
-        negative_keypoints = data["negativeKeypoints"]
+    # Create the 3D rectangle for the bounding box
+    y_padding = 0.1
+    c1 = np.array([x_left, y_left + y_padding, z_left])
+    c2 = np.array([x_right, y_right - y_padding, z_right])
+    c3 = np.array([x_left, y_left + y_padding, 0.05])
+    c4 = np.array([x_right, y_right - y_padding, 0.05])
 
-        x1 = bbox["x1"]
-        y1 = bbox["y1"]
-        x2 = bbox["x2"]
-        y2 = bbox["y2"]
+    # Generate all corners
+    corners_3d = np.array([c1, c2, c3, c4])
 
-        if "outlierThreshold" in data and data["outlierThreshold"] is not None:
-            outlier_threshold = float(data["outlierThreshold"])
-        else:
-            outlier_threshold = 0.5
+    X_C_W = np.linalg.inv(extrinsics)
+    projected_corners = project_points_to_image_plane(transform_points(X_C_W, corners_3d), intrinsics).squeeze()
 
-        scene_name = data["sceneName"]
+    # Get the 2D bounding box
+    u_min = min(u for u, _ in projected_corners)
+    v_min = min(v for _, v in projected_corners)
+    u_max = max(u for u, _ in projected_corners)
+    v_max = max(v for _, v in projected_corners)
 
-        image = cv2.imread(f"{scenes_directory}/{scene_name}/observation_result/image_left.png")
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-
-        predictor.set_image(image)
-
-        input_box = np.array([[x1, y1, x2, y2]])
-
-        input_point = []
-        input_label = []
-
-        for point in positive_keypoints:
-            input_point.append(point)
-            input_label.append(1)
-
-        for point in negative_keypoints:
-            input_point.append(point)
-            input_label.append(0)
-
-        logger.info(
-            f"Predicting for scene {scene_name} with box {input_box} and input points {input_point} and labels {input_label}"
-        )
-
-        masks, _, _ = predictor.predict(
-            point_coords=np.array(input_point) if len(input_point) > 0 else None,
-            point_labels=np.array(input_label) if len(input_label) > 0 else None,
-            box=input_box[None, :],
-            multimask_output=False,
-        )
-
-        mask = masks[0]
-
-        depth_map_path = f"{scenes_directory}/{scene_name}/observation_result/depth_map.tiff"
-        depth_map = cv2.imread(depth_map_path, cv2.IMREAD_UNCHANGED)
-
-        depth_map = np.nan_to_num(depth_map, nan=0, posinf=0, neginf=0)
-
-        masked_depth_map = np.where(mask > 0, depth_map, 0)
-        masked_values = masked_depth_map[mask > 0]
-
-        mean = np.mean(masked_values)
-
-        lower_bound = mean - outlier_threshold
-        upper_bound = mean + outlier_threshold
-
-        masked_depth_map = np.where(
-            (masked_depth_map > lower_bound) & (masked_depth_map < upper_bound), masked_depth_map, 0
-        )
-
-        x, y = np.meshgrid(np.arange(masked_depth_map.shape[1]), np.arange(masked_depth_map.shape[0]))
-
-        intrinsics = json.load(open(f"{scenes_directory}/{scene_name}/observation_result/camera_intrinsics.json"))
-
-        cx = intrinsics["principal_point_in_pixels"]["cx"]
-        cy = intrinsics["principal_point_in_pixels"]["cy"]
-
-        fx = intrinsics["focal_lengths_in_pixels"]["fx"]
-        fy = intrinsics["focal_lengths_in_pixels"]["fy"]
-
-        X1 = (x - 0.5 - cx) * masked_depth_map / fx
-        Y1 = (y - 0.5 - cy) * masked_depth_map / fy
-        X2 = (x + 0.5 - cx) * masked_depth_map / fx
-        Y2 = (y + 0.5 - cy) * masked_depth_map / fy
-
-        pixel_areas = np.abs((X2 - X1) * (Y2 - Y1))
-
-        mask_image_path = f"{scenes_directory}/{scene_name}/observation_result/mask.png"
-
-        cv2.imwrite(mask_image_path, ((pixel_areas > 0) * 255).astype(np.uint8))
-
-        coverage = np.nan_to_num(np.sum(pixel_areas), nan=0, posinf=0, neginf=0)
-
-        json.dump(
-            {
-                "bbox": {
-                    "x1": x1,
-                    "y1": y1,
-                    "x2": x2,
-                    "y2": y2,
-                },
-                "positiveKeypoints": positive_keypoints,
-                "negativeKeypoints": negative_keypoints,
-                "outlierThreshold": outlier_threshold,
-                "coverage": coverage,
-            },
-            open(f"{scenes_directory}/{scene_name}/observation_result/result.json", "w"),
-        )
-
-        print("Data processed")
-        if manual:
-            ack.put(scene_name)
+    return u_min, v_min, u_max, v_max
 
 
 def get_last_modified_time(directory):
@@ -311,85 +217,17 @@ def create_app(scenes_directory, q, ack, queued_scenes):  # noqa C901
         scenes_info = []
 
         for scene_name in scene_names:
-            dataset_path = os.path.join(scenes_directory, scene_name)
-            last_modified_time = get_last_modified_time(dataset_path)
+            sample_dir = os.path.join(scenes_directory, scene_name)
+            last_modified_time = get_last_modified_time(sample_dir)
 
-            result_exists = os.path.exists(f"{scenes_directory}/{scene_name}/observation_result/result.json")
-            result = None
+            result_path = f"{scenes_directory}/{scene_name}/observation_result/result.json"
+            result = json.load(open(result_path)) if os.path.exists(result_path) else None
 
-            if result_exists:
-                result = json.load(open(f"{scenes_directory}/{scene_name}/observation_result/result.json"))
-            elif scene_name not in queued_scenes:
-                intrinsics = json.load(
-                    open(f"{scenes_directory}/{scene_name}/observation_result/camera_intrinsics.json")
-                )
-                extrinsics = json.load(
-                    open(f"{scenes_directory}/{scene_name}/observation_result/camera_pose_in_world.json")
-                )
-                tcp_left = json.load(
-                    open(f"{scenes_directory}/{scene_name}/observation_result/arm_left_tcp_pose_in_world.json")
-                )
-                tcp_right = json.load(
-                    open(f"{scenes_directory}/{scene_name}/observation_result/arm_right_tcp_pose_in_world.json")
-                )
+            if result is None:
+                # Calculate initial bounding box
+                u_min, v_min, u_max, v_max = get_heuristic_cloth_bounding_box(sample_dir)
 
-                cx = intrinsics["principal_point_in_pixels"]["cx"]
-                cy = intrinsics["principal_point_in_pixels"]["cy"]
-
-                fx = intrinsics["focal_lengths_in_pixels"]["fx"]
-                fy = intrinsics["focal_lengths_in_pixels"]["fy"]
-
-                position = extrinsics["position_in_meters"]
-                rotation = extrinsics["rotation_euler_xyz_in_radians"]
-
-                # Convert rotation from Euler angles to rotation matrix
-                R = euler_to_mat(rotation["roll"], rotation["pitch"], rotation["yaw"])
-
-                # Convert position to numpy array
-                T = np.array([position["x"], position["y"], position["z"]])
-
-                tcp_left_position = tcp_left["position_in_meters"]
-                tcp_right_position = tcp_right["position_in_meters"]
-
-                x_left = tcp_left_position["x"]
-                y_left = tcp_left_position["y"]
-                z_left = tcp_left_position["z"]
-
-                x_right = tcp_right_position["x"]
-                y_right = tcp_right_position["y"]
-                z_right = tcp_right_position["z"]
-
-                # Create the 3D rectangle for the bounding box
-                y_padding = 0.1
-                c1 = np.array([x_left, y_left + y_padding, z_left])
-                c2 = np.array([x_right, y_right - y_padding, z_right])
-                c3 = np.array([x_left, y_left + y_padding, 0.05])
-                c4 = np.array([x_right, y_right - y_padding, 0.05])
-
-                # Generate all corners
-                corners = [c1, c2, c3, c4]
-
-                projected_corners = []
-                for corner in corners:
-                    # Convert world coordinates to camera coordinates
-                    X_cam = R.T @ (np.array(corner) - T)
-
-                    # Project to the image plane
-                    u = fx * (X_cam[0] / X_cam[2]) + cx
-                    v = fy * (X_cam[1] / X_cam[2]) + cy
-
-                    print("Projected corner", u, v, "for corner", corner)
-
-                    projected_corners.append((u, v))
-
-                # Get the 2D bounding box
-                u_min = min(u for u, _ in projected_corners)
-                v_min = min(v for _, v in projected_corners)
-                u_max = max(u for u, _ in projected_corners)
-                v_max = max(v for _, v in projected_corners)
-
-                print("Queueing unannotated scene", scene_name, "with bounding box", u_min, v_min, u_max, v_max)
-
+                # Queue the scene if it has not been segmented yet
                 q.put(
                     {
                         "sceneName": scene_name,
